IntroScreen State Repository Pattern implementiert
- Für IntroScreen State das Repository Pattern implementiert.
