+++ lib/models/intro_screen_state/intro_screen_state_interface.dart
@@ -1,0 +1,5 @@
+abstract class IntroScreenStateInterface {
+  void init();
+  Future<bool> getIntroScreenState();
+  void setIntroScreenState();
+}
